♻️ Refactor: Controller 역할에 따라서 분리
